Played with camera rotations and fixed a bug.
Added ability to rotate camera with Scroll Wheel of mouse, modified the movement scripts to make the character face the direction they are moving. Changed camera scripts so camera is no longer a child of objects (causing it to be destroyed when an object dies on its turn (due to bleed/recoil)).