Added hover event on UI, AOE range for skills, and dynamic grid generation.
Added an event system to UI for skills that displays information about the skill in a 'pop-up box' next to the skill. Also added a magenta 'net' around the targeted tile when using a skill with an AOE. Also, updated the grid generation so it ignores game objects with a certain layer, making the creation of levels at a later date pretty much 'drag-and-drop'.